backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "fastapi>=0.128.4",
    "passlib>=1.7.4",
    "psycopg>=3.3.3",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.46",
    "starlette>=0.52.1",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.0",
    "pydantic[email]>=2.12.5",
    "bcrypt==4.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
    "pytest>=9.0.3",
]
